requirements.txt
appdirs==1.4.4
bcrypt==3.1.7
blinker==1.6.2
cffi==1.15.1
click==8.1.7
cryptography==2.8
decorator==5.1.1
Deprecated==1.2.14
Fabric3==1.14.post1
Flask==2.3.3
greenlet==2.0.2
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
paramiko==2.12.0
pycparser==2.21
PyNaCl==1.3.0
pyparsing==3.1.1
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.7
wrapt==1.15.0
